+++ src/renderer/entities/pomodoro/index.ts
@@ -1,1 +1,2 @@
+export { pomodoroQueries } from './hooks/queries';
 export * from './api';
